ml/requirements.txt
# Blog Recommendation System Dependencies
# ========================================

# Database
pymongo==4.6.1

# ML & Embeddings
sentence-transformers==2.2.2
torch>=2.0.0

# Vector Database
pinecone-client==3.0.0

# Google AI (for embeddings)
google-generativeai==0.3.2

# Text Processing
markdown==3.5.1
beautifulsoup4==4.12.2
nltk==3.8.1

# Utilities
python-dotenv==1.0.0
numpy==1.26.3
tqdm==4.66.1
